### Description

This PR adds benchmarking for our parser against [pg-query-go](https://github.com/pganalyze/pg_query_go).

The benchmarks are run so - 
```
# Run with count=10 for good statistical confidence
go test -bench="BenchmarkMultigresParser$" -benchtime=180s -count=10 -run=^$ -benchmem > bench-multigres
go test -bench="BenchmarkPgQueryGo$" -benchtime=180s -count=10 -run=^$ -benchmem > bench-pgquery
sed 's/BenchmarkMultigresParser/BenchmarkParser/g' bench-multigres > bench-multigres-renamed
sed 's/BenchmarkPgQueryGo/BenchmarkParser/g' bench-pgquery > bench-pgquery-renamed
benchstat bench-multigres-renamed bench-pgquery-renamed
```

and here are the results - 
```
goos: darwin
goarch: arm64
pkg: github.com/multigres/multigres/go/parser
cpu: Apple M4
          │ bench-multigres-renamed │        bench-pgquery-renamed         │
          │         sec/op          │   sec/op     vs base                 │
Parser-10               144.5m ± 2%   365.7m ± 0%  +153.01% (p=0.000 n=10)

          │ bench-multigres-renamed │        bench-pgquery-renamed         │
          │          B/op           │     B/op      vs base                │
Parser-10             488.11Mi ± 0%   88.79Mi ± 0%  -81.81% (p=0.000 n=10)

          │ bench-multigres-renamed │        bench-pgquery-renamed        │
          │        allocs/op        │  allocs/op   vs base                │
Parser-10               2.388M ± 0%   1.782M ± 0%  -25.36% (p=0.000 n=10)
```

The memory and allocations I think can't be trusted because the go runtime doesn't account for the memory and allocations for the C plane that pq-query-go uses via CGO.

These are quite promising results. I also ran the benchmarks that are present in the [pg-query-go](https://github.com/pganalyze/pg_query_go) and here are the results - 

```
benchstat bench-multigres bench-pq-query
goos: darwin
goarch: arm64
pkg: github.com/pganalyze/pg_query_go/v6
cpu: Apple M4
                           │ bench-multigres │            bench-pq-query             │
                           │     sec/op      │    sec/op     vs base                 │
ParseSelect1-4                   1.556µ ± 1%    3.109µ ± 3%   +99.78% (p=0.000 n=10)
ParseSelect2-4                   3.173µ ± 0%   10.970µ ± 0%  +245.73% (p=0.000 n=10)
ParseCreateTable-4               7.653µ ± 0%   26.435µ ± 0%  +245.44% (p=0.000 n=10)
ParseSelect1Parallel-4          1632.5n ± 0%    916.0n ± 1%   -43.89% (p=0.000 n=10)
ParseSelect2Parallel-4           2.223µ ± 0%    3.191µ ± 2%   +43.54% (p=0.000 n=10)
ParseCreateTableParallel-4       3.924µ ± 0%    7.753µ ± 0%   +97.57% (p=0.000 n=10)
geomean                          2.852µ         5.228µ        +83.33%

                           │ bench-multigres │            bench-pq-query            │
                           │      B/op       │     B/op      vs base                │
ParseSelect1-4                   9786.0 ± 0%    1008.0 ± 0%  -89.70% (p=0.000 n=10)
ParseSelect2-4                 11.433Ki ± 0%   2.734Ki ± 0%  -76.08% (p=0.000 n=10)
ParseCreateTable-4             17.731Ki ± 0%   8.250Ki ± 0%  -53.47% (p=0.000 n=10)
ParseSelect1Parallel-4           9786.0 ± 0%    1008.0 ± 0%  -89.70% (p=0.000 n=10)
ParseSelect2Parallel-4         11.433Ki ± 0%   2.734Ki ± 0%  -76.08% (p=0.000 n=10)
ParseCreateTableParallel-4     17.730Ki ± 0%   8.250Ki ± 0%  -53.47% (p=0.000 n=10)
geomean                         12.47Ki        2.811Ki       -77.45%

                           │ bench-multigres │           bench-pq-query           │
                           │    allocs/op    │ allocs/op   vs base                │
ParseSelect1-4                    21.00 ± 0%   17.00 ± 0%  -19.05% (p=0.000 n=10)
ParseSelect2-4                    73.00 ± 0%   56.00 ± 0%  -23.29% (p=0.000 n=10)
ParseCreateTable-4                185.0 ± 0%   148.0 ± 0%  -20.00% (p=0.000 n=10)
ParseSelect1Parallel-4            21.00 ± 0%   17.00 ± 0%  -19.05% (p=0.000 n=10)
ParseSelect2Parallel-4            73.00 ± 0%   56.00 ± 0%  -23.29% (p=0.000 n=10)
ParseCreateTableParallel-4        185.0 ± 0%   148.0 ± 0%  -20.00% (p=0.000 n=10)
geomean                           65.70        52.04       -20.80%
```

The results are promising here too. Something that is baffling me is why the `ParseSelect1Parallel-4` slower in our code, while everything else is faster, especially because `ParseSelect1` is faster.